+++ app/src/scripts/components/header/Header.js
@@ -1,0 +1,9 @@
+import React from 'react';
+
+export default () => {
+	return (
+		<div className="header" style={{borderBottom: '1px solid #ccc'}}>
+			<header className="col-sm-2"><h1>Header</h1></header>
+		</div>
+	);
+}
