+++ app/src/scripts/components/pages/second/Second.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { Link } from 'react-router'
-import SearchInput from '../components/SearchInput';
+import SearchInput from './SearchInput';
 
 class Second extends React.Component {
 	render() {
@@ -12,8 +12,8 @@ class Second extends React.Component {
 				<SearchInput />
 
 				<ul>
-					<li><Link to="/First">Go back</Link></li>
-          <li><Link to="/Third">Go to view #3</Link></li>
+					<li><Link to="/">Go back</Link></li>
+          <li><Link to="Third">Go to view #3</Link></li>
         </ul>
 			</div>
 		)
