+++ app/src/scripts/main.js
@@ -1,10 +1,5 @@
 import React from 'react';
-import { render } from 'react-dom';
-import Router from 'react-router';
-import routes from './config/routes';
+import ReactDOM from 'react-dom';
+import App from './components/routes';
 
-render((
-  <Router>
-    { routes }
-  </Router>
-), document.getElementById('root'))
+ReactDOM.render(<App />, document.getElementById('root'));
