+++ package.json
@@ -32,6 +32,7 @@
     "watchify": "^3.6.1"
   },
   "dependencies": {
+    "flux": "^2.1.1",
     "react": "^0.14.3",
     "react-dom": "^0.14.3"
   }
